SQLの処理を一つのクラス（DAO）にまとめた
